spreadsheet2influx/requirements.txt
influxdb
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
pandas
python-dotenv
